refactor(api): move layered cache file and test
