+++ api/lib/infrastructure/caches/learning-content-cache.js
@@ -1,7 +1,7 @@
 import { config } from '../../../src/shared/config.js';
 import { DistributedCache } from '../../../src/shared/infrastructure/caches/DistributedCache.js';
 import { InMemoryCache } from '../../../src/shared/infrastructure/caches/InMemoryCache.js';
-import { LayeredCache } from './LayeredCache.js';
+import { LayeredCache } from '../../../src/shared/infrastructure/caches/LayeredCache.js';
 import { RedisCache } from './RedisCache.js';
 
 const LEARNING_CONTENT_CHANNEL = 'Learning content';
